refactor: Organize node reference utilities into dedicated directory
Move related files into node-references/ directory for better organization:
- node-reference.ts → node-references/formatting.ts
- node-reference-cleanup.ts → node-references/cleanup.ts
- detect-source-changes.ts → node-references/source-detection.ts

All exports are preserved and functionality remains unchanged.
