+++ packages/text-editor-utils/src/index.ts
@@ -1,6 +1,4 @@
 export * from "./extensions";
 export * from "./is-json-content";
 export * from "./json-content-to-text";
-export * from "./node-reference";
-export * from "./node-reference-cleanup";
-export * from "./detect-source-changes";
+export * from "./node-references";
