+++ packages/text-editor-utils/src/node-references/formatting.ts
