+++ packages/text-editor-utils/src/node-references/source-detection.ts
